tests/wpt/tests/css/css-break/overflow-clip-007.html

<!DOCTYPE html>
<link rel="author" title="Morten Stenshorne" href="mailto:mstensho@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-break-3/#fragmentation-model">
<link rel="help" href="https://www.w3.org/TR/css-overflow-3/#overflow-properties">
<meta name="assert" content="The spec says that when breakable content would overflow a fragmentainer in the block dimension, it breaks into the next container in its fragmentation context. If a container has clipped overflow, and the container ends before the fragmentation line, though, breakable content inside it can never overflow the fragmentainer, and as such, no additional fragmentainers should be generated.">
<div style="position:relative; columns:5; column-gap:20px; column-rule:20px solid green; column-fill:auto; width:180px; height:100px;">
  <div style="overflow:clip; height:300px;">
    <div style="height:300px; background:green;"></div>
    <div style="height:200px; background:red;"></div>
    <div id="target" style="break-before:column; height:10px; background:red;"></div>
  </div>
</div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(()=> {
      assert_equals(target.offsetLeft, 80);
      assert_equals(target.offsetTop, 300);
  }, "Overflowing and clipped content doesn't fragment");
</script>
